removed event bus & replaced w/ ObserveDispatch
